Parte 7 da aula 5 do curso "Android parte 3: refinando o projeto" em Kotlin. (Não compila)
